src/main/java/codeus/warmup/WarmupCalculator.java
package codeus.warmup;

public class WarmupCalculator {

    public static double calculate(String toCalculate) {
        String[] parts = toCalculate.split(" ");
        double operand1 = Double.parseDouble(parts[0]);
        String operator = parts[1];
        double operand2 = Double.parseDouble(parts[2]);

        switch (operator) {
            case "+":
                return operand1 + operand2;
            case "-":
                return operand1 - operand2;
            case "*":
                return operand1 * operand2;
            case "/":
                return operand1 / operand2;
            default:
                throw new IllegalArgumentException("Unknown operator: " + operator);
        }
    }

}